refactor: clean up A* implementation
