+++ lib/algorithms/astar/mod.rs
@@ -1,2 +1,3 @@
 pub mod heuristics;
 pub mod multiroom_distance_map;
+pub mod room_data_cache;
